In some cases the count query can take a lot of time and need an optimization.
For example if the the marketing activity list contains a lot of items (~1M) but only a ~10 different campaigns, postgres will ignore the index generation for campaign_id. This leads to query time above 30 seconds.

This fix will not count the items but will query for the first marketing activity item for the given campaign.